Rename io module and update main import
